refactor: Moved definition of DocumentNode
